\name{gibbs.mu.hier.gaussian}
\alias{gibbs.mu.hier.gaussian}
\title{Gibbs sample for mu in hierarchical gaussian model}
\usage{
  gibbs.mu.hier.gaussian(value, priors)
}
\arguments{
  \item{value}{}

  \item{priors}{}
}
\value{
  new parameter list
}
\description{
  Gibbs sample for mu in hierarchical gaussian model
}

