\name{gibbs.sigma.hier.gaussian}
\alias{gibbs.sigma.hier.gaussian}
\title{Gibbs sample sigma in hierarchical gaussian model}
\usage{
  gibbs.sigma.hier.gaussian(value, priors)
}
\arguments{
  \item{value}{}

  \item{priors}{}
}
\value{
  new parameter list
}
\description{
  Gibbs sample sigma in hierarchical gaussian model
}

