\name{loglikelihood.hier.gaussian}
\alias{loglikelihood.hier.gaussian}
\title{Example of a function computing the log likelihood.}
\usage{
  loglikelihood.hier.gaussian(y, theta, grad = FALSE)
}
\arguments{
  \item{y}{}

  \item{priors}{}

  \item{lp}{compute the log likelihood}

  \item{lgrad}{compute the log gradient}
}
\value{
  log posterior or log gradient
}
\description{
  Example of a function computing the log likelihood.
}

